Added to most GET operations to services Added testing for all

+++ src/main/java/com/version1/movies_and_shows_backend/models/CastId.java
@@ -5,15 +5,16 @@ import java.util.Objects;
 
 // Defining the composite primary key for the cast table
 public class CastId implements Serializable {
-    private Long personId;
+    private int personId;
     private String mediaId;
 
     public CastId() {} // required by JPA
 
-    public CastId(Long personId, String mediaId) {
+    public CastId(int personId, String mediaId) {
         this.personId = personId;
         this.mediaId = mediaId;
     }
+
 
     @Override
     public boolean equals(Object o) {
